refactor of default_url_on_new_port & using cli_factorys make_proxy
