+++ pytest_fixtures/component/smartproxy.py
@@ -2,7 +2,6 @@ import pytest
 from fauxfactory import gen_string
 
 from robottelo.cli.proxy import CapsuleTunnelError
-from robottelo.helpers import default_url_on_new_port
 
 
 @pytest.fixture(scope='session')
@@ -26,7 +25,7 @@ def module_fake_proxy(request, module_target_sat):
     args = {'name': gen_string(str_type='alpha')}
     newport = module_target_sat.get_available_capsule_port
     try:
-        with default_url_on_new_port(9090, newport) as url:
+        with module_target_sat.default_url_on_new_port(9090, newport) as url:
             args['url'] = url
             proxy = module_target_sat.api.SmartProxy(**args).create()
             yield proxy
