+++ robottelo/cli/factory.py
@@ -48,7 +48,6 @@ from robottelo.cli.operatingsys import OperatingSys
 from robottelo.cli.org import Org
 from robottelo.cli.partitiontable import PartitionTable
 from robottelo.cli.product import Product
-from robottelo.cli.proxy import CapsuleTunnelError
 from robottelo.cli.proxy import Proxy
 from robottelo.cli.realm import Realm
 from robottelo.cli.report_template import ReportTemplate
@@ -70,7 +69,6 @@ from robottelo.cli.virt_who_config import VirtWhoConfig
 from robottelo.config import settings
 from robottelo.datafactory import valid_cron_expressions
 from robottelo.decorators import cacheable
-from robottelo.helpers import default_url_on_new_port
 from robottelo.logging import logger
 
 
@@ -520,28 +518,6 @@ def make_product_wait(options=None, wait_for=5):
         if not product:
             raise err
     return product
-
-
-@cacheable
-def make_proxy(options=None):
-    """Creates a Proxy
-
-    :param options: Check options using `hammer proxy create --help` on satellite.
-
-    :returns Proxy object
-    """
-    args = {'name': gen_alphanumeric()}
-
-    if options is None or 'url' not in options:
-        newport = options['newport']
-        try:
-            with default_url_on_new_port(9090, newport) as url:
-                args['url'] = url
-                return create_object(Proxy, args, options)
-        except CapsuleTunnelError as err:
-            raise CLIFactoryError(f'Failed to create ssh tunnel: {err}')
-    args['url'] = options['url']
-    return create_object(Proxy, args, options)
 
 
 @cacheable
