+++ robottelo/helpers.py
@@ -1,5 +1,4 @@
 """Several helper methods and functions."""
-import contextlib
 import re
 from urllib.parse import urljoin  # noqa
 
@@ -8,13 +7,11 @@ from nailgun.config import ServerConfig
 
 from robottelo import ssh
 from robottelo.cli.base import CLIReturnCodeError
-from robottelo.cli.proxy import CapsuleTunnelError
 from robottelo.config import get_credentials
 from robottelo.config import get_url
 from robottelo.config import settings
 from robottelo.constants import PULP_PUBLISHED_YUM_REPOS_PATH
 from robottelo.errors import ProvisioningCheckError
-from robottelo.logging import logger
 
 
 def get_nailgun_config(user=None):
@@ -29,44 +26,6 @@ def get_nailgun_config(user=None):
     """
     creds = (user.login, user.passwd) if user else get_credentials()
     return ServerConfig(get_url(), creds, verify=False)
-
-
-@contextlib.contextmanager
-def default_url_on_new_port(oldport, newport):
-    """Creates context where the default capsule is forwarded on a new port
-
-    :param int oldport: Port to be forwarded.
-    :param int newport: New port to be used to forward `oldport`.
-
-    :return: A string containing the new capsule URL with port.
-    :rtype: str
-
-    """
-    domain = settings.server.hostname
-
-    client = ssh.get_client()
-    pre_ncat_procs = client.execute('pgrep ncat').stdout.splitlines()
-
-    with client.session.shell() as channel:
-        # if ncat isn't backgrounded, it prevents the channel from closing
-        command = f'ncat -kl -p {newport} -c "ncat {domain} {oldport}" &'
-        # broker 0.1.25 makes these debug messages redundant
-        logger.debug(f'Creating tunnel: {command}')
-        channel.send(command)
-        post_ncat_procs = client.execute('pgrep ncat').stdout.splitlines()
-        ncat_pid = set(post_ncat_procs).difference(set(pre_ncat_procs))
-        if not len(ncat_pid):
-            stderr = channel.get_exit_status()[1]
-            logger.debug(f'Tunnel failed: {stderr}')
-            # Something failed, so raise an exception.
-            raise CapsuleTunnelError(f'Starting ncat failed: {stderr}')
-        forward_url = f'https://{domain}:{newport}'
-        logger.debug(f'Yielding capsule forward port url: {forward_url}')
-        try:
-            yield forward_url
-        finally:
-            logger.debug(f'Killing ncat pid: {ncat_pid}')
-            client.execute(f'kill {ncat_pid.pop()}')
 
 
 class Storage:
