+++ robottelo/host_helpers/satellite_mixins.py
@@ -1,3 +1,4 @@
+import contextlib
 import random
 import re
 
@@ -8,6 +9,7 @@ from robottelo.cli.proxy import CapsuleTunnelError
 from robottelo.config import settings
 from robottelo.host_helpers.api_factory import APIFactory
 from robottelo.host_helpers.cli_factory import CLIFactory
+from robottelo.logging import logger
 
 
 class ContentInfo:
@@ -160,6 +162,38 @@ class SystemInfo:
             raise CapsuleTunnelError(
                 'Failed to create ssh tunnel: No more ports available for mapping'
             )
+
+    @contextlib.contextmanager
+    def default_url_on_new_port(self, oldport, newport):
+        """Creates context where the default capsule is forwarded on a new port
+
+        :param int oldport: Port to be forwarded.
+        :param int newport: New port to be used to forward `oldport`.
+
+        :return: A string containing the new capsule URL with port.
+        :rtype: str
+
+        """
+        pre_ncat_procs = self.execute('pgrep ncat').stdout.splitlines()
+        with self.session.shell() as channel:
+            # if ncat isn't backgrounded, it prevents the channel from closing
+            command = f'ncat -kl -p {newport} -c "ncat {self.hostname} {oldport}" &'
+            logger.debug(f'Creating tunnel: {command}')
+            channel.send(command)
+            post_ncat_procs = self.execute('pgrep ncat').stdout.splitlines()
+            ncat_pid = set(post_ncat_procs).difference(set(pre_ncat_procs))
+            if not len(ncat_pid):
+                stderr = channel.get_exit_status()[1]
+                logger.debug(f'Tunnel failed: {stderr}')
+                # Something failed, so raise an exception.
+                raise CapsuleTunnelError(f'Starting ncat failed: {stderr}')
+            forward_url = f'https://{self.hostname}:{newport}'
+            logger.debug(f'Yielding capsule forward port url: {forward_url}')
+            try:
+                yield forward_url
+            finally:
+                logger.debug(f'Killing ncat pid: {ncat_pid}')
+                self.execute(f'kill {ncat_pid.pop()}')
 
 
 class Factories:
