+++ tests/foreman/api/test_smartproxy.py
@@ -23,7 +23,6 @@ from requests import HTTPError
 
 from robottelo.datafactory import parametrized
 from robottelo.datafactory import valid_data_list
-from robottelo.helpers import default_url_on_new_port
 
 
 pytestmark = [pytest.mark.run_in_one_thread]
@@ -90,7 +89,7 @@ def test_positive_create_with_name(request, target_sat, name):
 
     """
     new_port = target_sat.get_available_capsule_port
-    with default_url_on_new_port(9090, new_port) as url:
+    with target_sat.default_url_on_new_port(9090, new_port) as url:
         proxy = _create_smart_proxy(request, target_sat, name=name, url=url)
         assert proxy.name == name
 
@@ -110,7 +109,7 @@ def test_positive_delete(target_sat):
     :BZ: 1398695
     """
     new_port = target_sat.get_available_capsule_port
-    with default_url_on_new_port(9090, new_port) as url:
+    with target_sat.default_url_on_new_port(9090, new_port) as url:
         proxy = entities.SmartProxy(url=url).create()
         proxy.delete()
     with pytest.raises(HTTPError):
@@ -130,7 +129,7 @@ def test_positive_update_name(request, target_sat):
 
     """
     new_port = target_sat.get_available_capsule_port
-    with default_url_on_new_port(9090, new_port) as url:
+    with target_sat.default_url_on_new_port(9090, new_port) as url:
         proxy = _create_smart_proxy(request, target_sat, url=url)
         for new_name in valid_data_list():
             proxy.name = new_name
@@ -152,11 +151,11 @@ def test_positive_update_url(request, target_sat):
     """
     # Create fake capsule
     port = target_sat.get_available_capsule_port
-    with default_url_on_new_port(9090, port) as url:
+    with target_sat.default_url_on_new_port(9090, port) as url:
         proxy = _create_smart_proxy(request, target_sat, url=url)
     # Open another tunnel to update url
     new_port = target_sat.get_available_capsule_port
-    with default_url_on_new_port(9090, new_port) as url:
+    with target_sat.default_url_on_new_port(9090, new_port) as url:
         proxy.url = url
         proxy = proxy.update(['url'])
         assert proxy.url == url
@@ -176,7 +175,7 @@ def test_positive_update_organization(request, target_sat):
     """
     organizations = [entities.Organization().create() for _ in range(2)]
     newport = target_sat.get_available_capsule_port
-    with default_url_on_new_port(9090, newport) as url:
+    with target_sat.default_url_on_new_port(9090, newport) as url:
         proxy = _create_smart_proxy(request, target_sat, url=url)
         proxy.organization = organizations
         proxy = proxy.update(['organization'])
@@ -197,7 +196,7 @@ def test_positive_update_location(request, target_sat):
     """
     locations = [entities.Location().create() for _ in range(2)]
     new_port = target_sat.get_available_capsule_port
-    with default_url_on_new_port(9090, new_port) as url:
+    with target_sat.default_url_on_new_port(9090, new_port) as url:
         proxy = _create_smart_proxy(request, target_sat, url=url)
         proxy.location = locations
         proxy = proxy.update(['location'])
@@ -223,7 +222,7 @@ def test_positive_refresh_features(request, target_sat):
 
     # get an available port for our fake capsule
     new_port = target_sat.get_available_capsule_port
-    with default_url_on_new_port(9090, new_port) as url:
+    with target_sat.default_url_on_new_port(9090, new_port) as url:
         proxy = _create_smart_proxy(request, target_sat, url=url)
         proxy.refresh()
 
@@ -243,7 +242,7 @@ def test_positive_import_puppet_classes(session_puppet_enabled_sat, puppet_proxy
     """
     with session_puppet_enabled_sat as puppet_sat:
         new_port = puppet_sat.get_available_capsule_port
-        with default_url_on_new_port(9090, new_port) as url:
+        with puppet_sat.default_url_on_new_port(9090, new_port) as url:
             proxy = entities.SmartProxy(url=url).create()
             result = proxy.import_puppetclasses()
             assert (
